+++ src/main/java/com/example/bandShop/service/ProductServise.java
@@ -4,8 +4,10 @@ import com.example.bandShop.entity.ProductEntity;
 import com.example.bandShop.exception.ProductAlreadyExistException;
 import com.example.bandShop.exception.ProductNotFoundedException;
 import com.example.bandShop.exception.SortOrderException;
+import com.example.bandShop.exception.UserNotFoundException;
 import com.example.bandShop.model.Product;
 import com.example.bandShop.model.ProductCard;
+import com.example.bandShop.model.User;
 import com.example.bandShop.repository.ProductRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -21,13 +23,16 @@ public class ProductServise {
     @Autowired
     private ProductRepo productRepo;
 
-    public String delete(String id)  {
+    public Product delete(String id) throws ProductNotFoundedException {
+
+        if(!productRepo.existsById(id))
+            throw  new ProductNotFoundedException("Проукт не найден");
         productRepo.deleteById(id);
-        return id;
+        return new Product();
     }
 
     public ProductEntity createProduct (ProductEntity product) throws ProductAlreadyExistException {
-        if (!productRepo.findById(product.getId()).isPresent())
+        if (productRepo.existsById(product.getId()))
             throw new ProductAlreadyExistException("Продукт уже существует");
         product.setStrorage(0);
         product.setReviews(new ArrayList<>());
@@ -35,16 +40,19 @@ public class ProductServise {
     }
 
     public Product addProduct (String id, int amount) throws ProductNotFoundedException {
-        ProductEntity product = productRepo.findById(id).get();
-        if (product == null)
+
+        if (!productRepo.existsById(id))
             throw new ProductNotFoundedException("Продукт не найден");
+        ProductEntity product = productRepo.findById(id).get();
         product.setStrorage(product.getStrorage()+amount);
         productRepo.save(product);
         return Product.toModel(product);
     }
 
     public Product updateProduct (ProductEntity product) throws ProductNotFoundedException {
-        if(!productRepo.findById(product.getId()).isPresent()){
+
+        if(productRepo.existsById(product.getId())){
+            product.setReviews(productRepo.findById(product.getId()).get().getReviews());
             productRepo.save(product);
             return Product.toModel(product);
         }
@@ -52,10 +60,9 @@ public class ProductServise {
     }
 
     public Product getOneProduct (String id) throws ProductNotFoundedException {
-        ProductEntity product = productRepo.findById(id).get();
-        if(product == null)
+        if(!productRepo.existsById(id))
             throw new ProductNotFoundedException("Продукт не найден");
-        return Product.toModel(product);
+        return Product.toModel(productRepo.findById(id).get());
     }
 
     public List<ProductCard> getProducts() {
@@ -79,9 +86,9 @@ public class ProductServise {
 
     public List<ProductCard> getProductsSortedByTitle(String order) throws SortOrderException {
         List<ProductCard> products = getProducts();
-        switch (order){
-            case "increase"->products.sort(Comparator.comparing(ProductCard::getTitle));
-            case "decrease"->products.sort(Comparator.comparing(ProductCard::getTitle).reversed());
+        switch (order.trim()){
+            case "increase"-> products.sort(Comparator.comparing(ProductCard::getTitle));
+            case "decrease"-> products.sort(Comparator.comparing(ProductCard::getTitle).reversed());
             default -> throw new SortOrderException("Не верный порядок сортировки");
         }
         return products;
@@ -112,9 +119,9 @@ public class ProductServise {
         List<ProductCard> products = getProducts();
         List<ProductCard> sortProducts = new ArrayList<>();
         for(ProductCard pc : products)
-            if(pc.getTitle().contains(request) || pc.getId().contains(request))
+            if(pc.getTitle().contains(request.trim()) || pc.getId().contains(request.trim()))
                 sortProducts.add(pc);
-        if(sortProducts.get(0) == null)
+        if(sortProducts.isEmpty())
             throw new ProductNotFoundedException("Продукт не найден");
         return sortProducts;
     }
