+++ src/main/java/com/example/bandShop/service/UserService.java
@@ -4,12 +4,14 @@ import com.example.bandShop.entity.CartEntity;
 import com.example.bandShop.entity.ProductEntity;
 import com.example.bandShop.entity.UserEntity;
 import com.example.bandShop.exception.PasswordUncorectException;
+import com.example.bandShop.exception.ProductNotFoundedException;
 import com.example.bandShop.exception.UserAlreadyExistException;
 import com.example.bandShop.exception.UserNotFoundException;
 import com.example.bandShop.model.OrdersHistory;
 import com.example.bandShop.model.PersonalInformation;
 import com.example.bandShop.model.User;
 import com.example.bandShop.model.WishList;
+import com.example.bandShop.repository.ProductRepo;
 import com.example.bandShop.repository.UserRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -22,33 +24,36 @@ public class UserService {
 
     @Autowired
     private UserRepo userRepo;
+    @Autowired
+    private ProductRepo productRepo;
 
     public UserEntity registration(UserEntity user) throws UserAlreadyExistException {
         UserEntity usere = userRepo.findByLogin(user.getLogin());
         if (usere != null){
             throw new UserAlreadyExistException("Пользователь с таким логином уже существует");
         }
-        usere.setCart(new CartEntity());
-        usere.setOrderHistory(new ArrayList<>());
-        usere.setReviews(new ArrayList<>());
-        usere.setWishlist(new ArrayList<>());
-        return userRepo.save(usere);
+        user.setCart(new CartEntity());
+        user.setOrderHistory(new ArrayList<>());
+        user.setReviews(new ArrayList<>());
+        user.setWishlist(new ArrayList<>());
+        return userRepo.save(user);
     }
 
     public User getOne(Integer id) throws UserNotFoundException {
-        UserEntity user = userRepo.findById(id).get();
-        if(user == null)
+        if(!userRepo.existsById(id))
             throw new UserNotFoundException("Пользователь не был найден");
-        return User.toModel(user);
+        return User.toModel(userRepo.findById(id).get());
     }
 
-    public Integer delete(Integer id)  {
+    public User delete(Integer id) throws UserNotFoundException {
+        if(!userRepo.existsById(id))
+            throw  new UserNotFoundException("Пользователь не найден");
         userRepo.deleteById(id);
-        return id;
+        return new User();
     }
 
     public User updateUser (UserEntity user) throws UserNotFoundException, UserAlreadyExistException {
-        if(!userRepo.findById(user.getId()).isPresent()){
+        if(userRepo.findById(user.getId()).isPresent()){
             if(userRepo.findByLogin(user.getLogin()) != null && userRepo.findByLogin(user.getLogin()) != userRepo.findById(user.getId()).get())
                 throw new UserAlreadyExistException("Логин занят");
             userRepo.save(user);
@@ -59,7 +64,7 @@ public class UserService {
 
     public User enter(UserEntity user) throws UserNotFoundException, PasswordUncorectException {
         UserEntity usere = userRepo.findByLogin(user.getLogin());
-        if(user == null)
+        if(usere == null)
             throw new UserNotFoundException("Пользователь не был найден");
         if(!user.getPassword().equals(usere.getPassword()))
             throw new PasswordUncorectException("Не верный пароль");
@@ -75,34 +80,52 @@ public class UserService {
         return users;
     }
 
-    public WishList addToWishList(ProductEntity product, int user_id) throws UserNotFoundException {
-        UserEntity user = userRepo.findById(user_id).get();
-        if(user == null)
+    public WishList addToWishList(String product_id, int user_id) throws UserNotFoundException, ProductNotFoundedException {
+
+        if(!userRepo.existsById(user_id))
             throw new UserNotFoundException("Пользователь не был найден");
-        user.getWishlist().add(product);
+        UserEntity user = userRepo.findById(user_id).get();
+        if(!productRepo.existsById(product_id))
+            throw new ProductNotFoundedException("Продукт не найден");
+        user.getWishlist().add(productRepo.findById(product_id).get());
+        userRepo.save(user);
+        return WishList.toModel(user);
+    }
+
+    public WishList dellFromWishList(String product_id, int user_id) throws UserNotFoundException, ProductNotFoundedException {
+
+        if(!userRepo.existsById(user_id))
+            throw new UserNotFoundException("Пользователь не был найден");
+        UserEntity user = userRepo.findById(user_id).get();
+        if(!productRepo.existsById(product_id))
+            throw new ProductNotFoundedException("Продукт не найден");
+        user.getWishlist().remove(productRepo.findById(product_id).get());
         userRepo.save(user);
         return WishList.toModel(user);
     }
 
     public PersonalInformation updatePerosnalInfo(UserEntity user) throws UserNotFoundException {
-        if(!userRepo.findById(user.getId()).isPresent()){
-            userRepo.save(user);
+        if(userRepo.existsById(user.getId())){
+            UserEntity usere = userRepo.findById(user.getId()).get();
+            usere.setBirthday(user.getBirthday());
+            usere.setFirstName(user.getFirstName());
+            usere.setLastName(user.getLastName());
+            usere.setMobile(user.getMobile());
+            userRepo.save(usere);
             return PersonalInformation.toModel(user);
         }
         throw new UserNotFoundException("Пользователь не найден");
     }
 
     public WishList getWishList(int user_id) throws UserNotFoundException {
-        UserEntity user = userRepo.findById(user_id).get();
-        if(user == null)
+        if(!userRepo.existsById(user_id))
             throw new UserNotFoundException("Пользователь не был найден");
-        return WishList.toModel(user);
+        return WishList.toModel(userRepo.findById(user_id).get());
     }
 
     public OrdersHistory getOrderHistory(int user_id) throws UserNotFoundException {
-        UserEntity user = userRepo.findById(user_id).get();
-        if(user == null)
+        if(!userRepo.existsById(user_id))
             throw new UserNotFoundException("Пользователь не был найден");
-        return OrdersHistory.toModel(user);
+        return OrdersHistory.toModel(userRepo.findById(user_id).get());
     }
 }
